requirements.txt
python-telegram-bot==20.*
yt-dlp
flask
requests